added new companionLayout component
